+++ tests/TestAsset/configs/without-params/config/web.php
@@ -1,0 +1,9 @@
+<?php
+
+declare(strict_types=1);
+
+/** @var array $params */
+
+return [
+    'root-web-key' => $params['root-params-key'] ?? 42,
+];